Create module ospd.error and ospd.xml Move OSPDError class from ospd.ospd to its own module. Move simple_response_str() and get_result_xml() from ospd.ospd to its own module. This allows to import the modules from other files, avoiding an import loop. Update tests accordingly.
